feat: user and contact relationship established

+++ controllers/contactController.js
@@ -1,28 +1,30 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
+
 //================ Public functions ================//
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access Public
+//@access Private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find({});
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 });
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access Public
+//@access Private
 
-const createContact = async (req, res) => {
+const createContact = asyncHandler(async (req, res) => {
   console.log("req.body:", req.body);
   const { name, surname, email, phone, country, city } = req.body;
   if (!name || !surname || !email || !phone || !country || !city) {
     return res.status(400).json({ message: "Please enter all fields!" });
   }
   const contact = await Contact.create({
+    user_id: req.user.id,
     name,
     surname,
     email,
@@ -31,21 +33,21 @@ const createContact = async (req, res) => {
     city,
   });
   res.status(201).json({ message: `Contact '${name} ${surname}' created successfully ✅` });
-};
+});
 
 //================ Private functions ================//
 
 //@desc Get single contact
 //@route GET /api/contacts/:id
-//@access Public
+//@access Private
 
-const getContact = async (req, res) => {
+const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     return res.status(404).json({ message: `Contact with id ${req.params.id} not found!` });
   } 
   res.status(200).json(contact);
-};
+});
 
 
 
@@ -53,35 +55,43 @@ const getContact = async (req, res) => {
 //@route PUT /api/contacts/:id
 //@access Private
 
-const updateContact = async (req, res) => {
-  const { name, surname, email, phone, country, city } = req.body;
-  if (!name || !surname || !email || !phone || !country || !city) {
-    return res.status(400).json({ message: "Please enter all fields!" });
+const updateContact = asyncHandler(async (req, res) => {
+  const contact = await Contact.findById(req.params.id);
+  if (!contact) {
+    res.status(404);
+    throw new Error(`Contact with id ${req.params.id} not found!`);
+  }
+  if (contact.user_id.toString() !== req.user.id.toString()) {
+    res.status(403);
+    throw new Error("User do not have permission to UPDATE other user contacts");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
-    { name, surname, email, phone, country, city } || req.body,
-    { new: true }
+    req.body,
+    { new: true, runValidators: true }
   );
-  if (!updatedContact) {
-    return res.status(404).json({ message: `Contact with id ${req.params.id} not found!` });
-  }
-  res.status(200).json(updatedContact);
-};
+  res.status(200).json({ message: `Contact '${updatedContact.name} ${updatedContact.surname}' updated successfully...` });  
+});
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
 //@access Private
 
-const deleteContact = async (req, res) => {
+const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   const deletedContact = await Contact.findByIdAndDelete(req.params.id);
   if (!contact) {
     return res.status(404).json({ message: `Contact with id ${req.params.id} not found!` });
   }
+   if (contact.user_id.toString() !== req.user.id.toString()) {
+     res.status(403);
+     throw new Error(
+       "User do not have permission to DELETE other user contacts"
+     );
+   }
   await Contact.findByIdAndDelete(req.params.id);
-  res.status(200).json({ message: `Contact ${deletedContact.name} ${deletedContact.surname} deleted successfully...` });
-};
+  res.status(200).json({ message: `Contact '${deletedContact.name} ${deletedContact.surname}' deleted successfully...` });
+});
 
 module.exports = {
   getContacts,
